extract dataclasses for industry2030
