+++ generatorcore/industry2030/__init__.py
@@ -1,471 +1,7 @@
-from dataclasses import dataclass, field
-
 from .. import industry2018
 from ..inputs import Inputs
 from ..utils import div
-
-
-@dataclass
-class Vars0:
-    # Used by g
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_com: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-
-
-@dataclass
-class Vars1:
-    # Used by g_consult
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_com: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    emplo_existing: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars2:
-    # Used by i
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_com: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-
-
-@dataclass
-class Vars3:
-    # Used by p
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_biomass: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emethan: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    demand_heatnet: float = None  # type: ignore
-    demand_hydrogen: float = None  # type: ignore
-    emplo_existing: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-
-
-@dataclass
-class Vars4:
-    # Used by p_miner
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_biomass: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emethan: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_hydrogen: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-
-
-@dataclass
-class Vars5:
-    # Used by p_miner_cement, p_miner_chalk, p_chem_basic, p_chem_other
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_t: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_production_based_per_t: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_change: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emethan: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars6:
-    # Used by p_miner_glas, p_chem_ammonia, p_metal_steel_secondary
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_t: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_production_based_per_t: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_change: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars7:
-    # Used by p_miner_ceram, p_metal_nonfe
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_t: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_production_based_per_t: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_biomass: float = None  # type: ignore
-    demand_change: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_hydrogen: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars8:
-    # Used by p_chem
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emethan: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-
-
-@dataclass
-class Vars9:
-    # Used by p_metal
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_biomass: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_hydrogen: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-
-
-@dataclass
-class Vars10:
-    # Used by p_metal_steel
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-
-
-@dataclass
-class Vars11:
-    # Used by p_metal_steel_primary
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_t: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_production_based_per_t: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_change: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_hydrogen: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars12:
-    # Used by p_other
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emethan: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_heatnet: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-
-
-@dataclass
-class Vars13:
-    # Used by p_other_paper, p_other_food
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_t: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_production_based_per_t: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_change: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_heatnet: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars14:
-    # Used by p_other_further
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_MWh: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_production_based_per_MWh: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_change: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emethan: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_heatnet: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    prod_volume: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars15:
-    # Used by p_other_2efgh
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_production_based_per_MWh: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_outside: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_outside: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars16:
-    # Used by s, s_fossil_gas, s_fossil_coal, s_fossil_diesel, s_fossil_fueloil, s_fossil_lpg, s_fossil_opetpro, s_fossil_ofossil, s_renew, s_renew_hydrogen, s_renew_emethan, s_renew_biomass, s_renew_heatnet, s_renew_heatpump, s_renew_solarth, s_renew_elec
-    energy: float = None  # type: ignore
-
-
-@dataclass
-class Vars17:
-    # Used by s_fossil
-    pass
-
-
-@dataclass
-class I30:
-    g: Vars0 = field(default_factory=Vars0)
-    g_consult: Vars1 = field(default_factory=Vars1)
-    i: Vars2 = field(default_factory=Vars2)
-    p: Vars3 = field(default_factory=Vars3)
-    p_miner: Vars4 = field(default_factory=Vars4)
-    p_miner_cement: Vars5 = field(default_factory=Vars5)
-    p_miner_chalk: Vars5 = field(default_factory=Vars5)
-    p_miner_glas: Vars6 = field(default_factory=Vars6)
-    p_miner_ceram: Vars7 = field(default_factory=Vars7)
-    p_chem: Vars8 = field(default_factory=Vars8)
-    p_chem_basic: Vars5 = field(default_factory=Vars5)
-    p_chem_ammonia: Vars6 = field(default_factory=Vars6)
-    p_chem_other: Vars5 = field(default_factory=Vars5)
-    p_metal: Vars9 = field(default_factory=Vars9)
-    p_metal_steel: Vars10 = field(default_factory=Vars10)
-    p_metal_steel_primary: Vars11 = field(default_factory=Vars11)
-    p_metal_steel_secondary: Vars6 = field(default_factory=Vars6)
-    p_metal_nonfe: Vars7 = field(default_factory=Vars7)
-    p_other: Vars12 = field(default_factory=Vars12)
-    p_other_paper: Vars13 = field(default_factory=Vars13)
-    p_other_food: Vars13 = field(default_factory=Vars13)
-    p_other_further: Vars14 = field(default_factory=Vars14)
-    p_other_2efgh: Vars15 = field(default_factory=Vars15)
-    s: Vars16 = field(default_factory=Vars16)
-    s_fossil: Vars17 = field(default_factory=Vars17)
-    s_fossil_gas: Vars16 = field(default_factory=Vars16)
-    s_fossil_coal: Vars16 = field(default_factory=Vars16)
-    s_fossil_diesel: Vars16 = field(default_factory=Vars16)
-    s_fossil_fueloil: Vars16 = field(default_factory=Vars16)
-    s_fossil_lpg: Vars16 = field(default_factory=Vars16)
-    s_fossil_opetpro: Vars16 = field(default_factory=Vars16)
-    s_fossil_ofossil: Vars16 = field(default_factory=Vars16)
-    s_renew: Vars16 = field(default_factory=Vars16)
-    s_renew_hydrogen: Vars16 = field(default_factory=Vars16)
-    s_renew_emethan: Vars16 = field(default_factory=Vars16)
-    s_renew_biomass: Vars16 = field(default_factory=Vars16)
-    s_renew_heatnet: Vars16 = field(default_factory=Vars16)
-    s_renew_heatpump: Vars16 = field(default_factory=Vars16)
-    s_renew_solarth: Vars16 = field(default_factory=Vars16)
-    s_renew_elec: Vars16 = field(default_factory=Vars16)
+from .i30 import I30
 
 
 def calc(inputs: Inputs, *, i18: industry2018.I18) -> I30:
