+++ generatorcore/industry2030/i30.py
@@ -1,0 +1,66 @@
+from dataclasses import dataclass, field
+
+from .dataclasses import (
+    Vars0,
+    Vars1,
+    Vars2,
+    Vars3,
+    Vars4,
+    Vars5,
+    Vars6,
+    Vars7,
+    Vars8,
+    Vars9,
+    Vars10,
+    Vars11,
+    Vars12,
+    Vars13,
+    Vars14,
+    Vars15,
+    Vars16,
+    Vars17,
+)
+
+
+@dataclass
+class I30:
+    g: Vars0 = field(default_factory=Vars0)
+    g_consult: Vars1 = field(default_factory=Vars1)
+    i: Vars2 = field(default_factory=Vars2)
+    p: Vars3 = field(default_factory=Vars3)
+    p_miner: Vars4 = field(default_factory=Vars4)
+    p_miner_cement: Vars5 = field(default_factory=Vars5)
+    p_miner_chalk: Vars5 = field(default_factory=Vars5)
+    p_miner_glas: Vars6 = field(default_factory=Vars6)
+    p_miner_ceram: Vars7 = field(default_factory=Vars7)
+    p_chem: Vars8 = field(default_factory=Vars8)
+    p_chem_basic: Vars5 = field(default_factory=Vars5)
+    p_chem_ammonia: Vars6 = field(default_factory=Vars6)
+    p_chem_other: Vars5 = field(default_factory=Vars5)
+    p_metal: Vars9 = field(default_factory=Vars9)
+    p_metal_steel: Vars10 = field(default_factory=Vars10)
+    p_metal_steel_primary: Vars11 = field(default_factory=Vars11)
+    p_metal_steel_secondary: Vars6 = field(default_factory=Vars6)
+    p_metal_nonfe: Vars7 = field(default_factory=Vars7)
+    p_other: Vars12 = field(default_factory=Vars12)
+    p_other_paper: Vars13 = field(default_factory=Vars13)
+    p_other_food: Vars13 = field(default_factory=Vars13)
+    p_other_further: Vars14 = field(default_factory=Vars14)
+    p_other_2efgh: Vars15 = field(default_factory=Vars15)
+    s: Vars16 = field(default_factory=Vars16)
+    s_fossil: Vars17 = field(default_factory=Vars17)
+    s_fossil_gas: Vars16 = field(default_factory=Vars16)
+    s_fossil_coal: Vars16 = field(default_factory=Vars16)
+    s_fossil_diesel: Vars16 = field(default_factory=Vars16)
+    s_fossil_fueloil: Vars16 = field(default_factory=Vars16)
+    s_fossil_lpg: Vars16 = field(default_factory=Vars16)
+    s_fossil_opetpro: Vars16 = field(default_factory=Vars16)
+    s_fossil_ofossil: Vars16 = field(default_factory=Vars16)
+    s_renew: Vars16 = field(default_factory=Vars16)
+    s_renew_hydrogen: Vars16 = field(default_factory=Vars16)
+    s_renew_emethan: Vars16 = field(default_factory=Vars16)
+    s_renew_biomass: Vars16 = field(default_factory=Vars16)
+    s_renew_heatnet: Vars16 = field(default_factory=Vars16)
+    s_renew_heatpump: Vars16 = field(default_factory=Vars16)
+    s_renew_solarth: Vars16 = field(default_factory=Vars16)
+    s_renew_elec: Vars16 = field(default_factory=Vars16)
